@@ -16,26 +16,26 @@
         "chartist-plugin-legend": "^0.6.2",
         "chartist-plugin-pointlabels": "^0.0.6",
         "chartist-plugin-tooltips-updated": "^0.1.2",
-        "codemirror": "^5.51.0",
+        "codemirror": "^5.65.9",
         "cytoscape": "^3.13.0",
         "cytoscape-canvas": "^3.0.1",
         "cytoscape-context-menus": "^4.0.0",
         "cytoscape-dagre": "^2.2.1",
         "cytoscape-grid-guide": "^2.3.0",
         "diff-match-patch": "^1.0.4",
-        "file-type": "^14.1.3",
+        "file-type": "^17.1.6",
         "fittext.js": "^1.2.0",
         "flatpickr": "^4.6.3",
         "fuzzy": "^0.1.3",
-        "gettext.js": "^1.0.0",
+        "gettext.js": "^1.2.0",
         "gridstack": "^0.6.0",
         "html5sortable": "^0.9.18",
         "jquery": "^3.4.1",
         "jquery-migrate": "^3.1.0",
         "jquery-mousewheel": "^3.1.13",
         "jquery-prettytextdiff": "^1.0.4",
-        "jquery-ui": "^1.11.4",
-        "jquery-ui-dist": "^1.11.4",
+        "jquery-ui": "^1.13.2",
+        "jquery-ui-dist": "^1.13.2",
         "jquery.autogrow-textarea": "^0.4.1",
         "jquery.rateit": "^1.1.3",
         "jstree": "^3.3.8",
@@ -46,13 +46,13 @@
         "leaflet.markercluster": "^1.4.1",
         "lodash": "^4.17.21",
         "photoswipe": "^4.1.3",
-        "prismjs": "^1.23.0",
+        "prismjs": "^1.29.0",
         "qtip2": "^3.0.3",
         "rrule": "^2.6.3",
         "select2": "^4.0.13",
         "spin.js": "^4.1.0",
-        "tinymce": "^4.9.8",
-        "tinymce-i18n": "^20.4.4",
+        "tinymce": "^5.10.6",
+        "tinymce-i18n": "^22.10.18",
         "unorm": "^1.6.0"
     },
     "scripts": {
@@ -64,7 +64,7 @@
         "clean-webpack-plugin": "^4.0.0",
         "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.7.1",
-        "eslint": "^7.1.0",
+        "eslint": "^7.32.0",
         "glob": "^8.0.3",
         "mini-css-extract-plugin": "^2.6.1",
         "os-browserify": "^0.3.0",
