@@ -29,4 +29,16 @@
  * ---------------------------------------------------------------------
  */
 
-window.fileType = require('file-type/browser');
+import {
+   fileTypeFromBuffer,
+   fileTypeFromStream,
+   fileTypeFromTokenizer,
+   fileTypeStream,
+} from 'file-type';
+
+window.fileType = {
+   fromBuffer: fileTypeFromBuffer,
+   fromStream: fileTypeFromStream,
+   fromTokenizer: fileTypeFromTokenizer,
+   stream: fileTypeStream,
+};
