@@ -32,23 +32,23 @@
 // 'tinymce' and 'tinyMCE' objects have to be declared in global scope
 window.tinymce = window.tinyMCE = require('tinymce/tinymce');
 
-// Base theme
-require('tinymce/themes/modern/theme');
+// Default icons
+require('tinymce/icons/default/icons');
+
+// Base theme / skin
+require('tinymce/themes/silver/theme');
 
 // Used plugins
-require('tinymce/plugins/table');
+require('tinymce/plugins/autoresize');
+require('tinymce/plugins/code');
+require('tinymce/plugins/colorpicker');
 require('tinymce/plugins/directionality');
+require('tinymce/plugins/fullscreen');
+require('tinymce/plugins/image');
+require('tinymce/plugins/link');
+require('tinymce/plugins/lists');
+require('tinymce/plugins/paste');
 require('tinymce/plugins/searchreplace');
 require('tinymce/plugins/tabfocus');
-require('tinymce/plugins/autoresize');
-require('tinymce/plugins/link');
-require('tinymce/plugins/image');
-require('tinymce/plugins/paste');
-require('tinymce/plugins/code');
-require('tinymce/plugins/fullscreen');
+require('tinymce/plugins/table');
 require('tinymce/plugins/textcolor');
-require('tinymce/plugins/colorpicker');
-require('tinymce/plugins/lists');
-
-// Custom plugins
-require('../tiny_mce/custom_plugins/stickytoolbar/plugin')
